libmix: correct the type of buffers in structure _VASurfaceAttributeTPI
BZ: 172915

correct the type of buffers in structure _VASurfaceAttributeTPI

+++ imagedecoder/JPEGDecoder_img.cpp
@@ -109,7 +109,7 @@ JpegDecodeStatus JpegDecoder::createSurfaceGralloc(int width, int height, uint32
     attrib_tpi.width = width;
     attrib_tpi.height = height;
     attrib_tpi.type = VAExternalMemoryAndroidGrallocBuffer;
-    attrib_tpi.buffers = (uint32_t*)&handle;
+    attrib_tpi.buffers = (unsigned long *)&handle;
 
     st = vaCreateSurfacesWithAttribute(
         mDisplay,
